+++ src/restSpa/routes/utils.nim
@@ -1,6 +1,7 @@
 from std/logging import nil
 
 from std/json import parseJson, `{}=`, `%`, newJObject
+
 from std/strtabs import keys
 
 import restSpa/config
@@ -29,6 +30,7 @@ proc hasKey*(session: var Session; key: string): bool =
 using
   ctx: Context
   node: JsonNode
+  body: untyped
 
 proc setContentJsonHeader*(ctx) =
   ## Set the response content-type to json
@@ -60,6 +62,7 @@ template withParams*(ctx; mergeGet = false; bodyCode: untyped) =
         for key, val in ctx.request.formParams.data:
           # echo val.params
           node{key} = %val.body
+        error = false
     of "application/xml":
       when autoXmlParsing:
         {.fatal: "XML parsing not implemented".}
@@ -87,36 +90,54 @@ template respSucJson*(json: string; code = Http200) =
   ## Send a error message in a JSON to client
   respJson({"kind": "json", "json": json, "error": false}, code)
 
-from std/strutils import `%`
+from std/strutils import `%`, join
 
-const ifContainsDefaultErr* = "No $1 provided"
+template ifNoError(body) =
+  ## If `error` is false execute the body, else shows a error
+  if error: respErr ifContainsInvalidReq
+  else: body
+
 template ifContains*(
   node;
-  fields: openArray[string];
-  errorMsg: string;
-  body: untyped
+  all: openArray[string];
+  body
 ) =
-  ## Checks if the json have errors and specific fields, if not exists,
-  ## send the error message.
-  ##
-  ## If all is good, the body is executed
-  if error:
-    respJson({"message": "Invalid request", "error": true}, Http400)
-  else:
-    var haveAll = true
-    for field in fields:
-      if not node.hasKey field:
-        respErr errorMsg % field
-        haveAll = false
-        break
-    if haveAll:
+  ## Checks if the json have errors and `all` fields, if not exists,
+  ## send the error message and not executes the body
+  ifNoError:
+    var havent = @all
+    for field in all:
+      if node.hasKey field:
+        havent.delete havent.find field
+    if havent.len > 0:
+      respErr ifContainsAllErr % havent.join ", "
+    else:
       body
+
+template ifContains*(
+  node;
+  atLeast: openArray[string];
+  body
+) =
+  ## Checks if the json have errors and `atLeast` fields, if not exists,
+  ## send the error message and not executes the body
+  ifNoError:
+    if atLeast.len > 0:
+      var have = false
+      for field in atLeast:
+        if node.hasKey field:
+          have = true
+      if not have:
+        respErr ifContainsAtLeastErr % atLeast.join ", "
+      else:
+        body
+
 
 template forceHttpMethod*(ctx; httpMethod: HttpMethod) =
   ## Forces an specific HTTP method
   ##
   ## Useful just in development
-  doAssert ctx.request.reqMethod == httpMethod
+  assert ctx.request.reqMethod == httpMethod
 
 proc getSession*(ctx; name: string; default = ""): string =
   ## Tries to get a session value
@@ -126,7 +147,7 @@ proc getSession*(ctx; name: string; default = ""): string =
   except: discard
 
 
-template ifLogin*(ctx; loggedIn = true; body: untyped) =
+template ifLogin*(ctx; loggedIn = true; body) =
   ## Runs the code if user have login
   var check = ctx.getSession(sess_username).len > 0
   if not loggedIn:
@@ -135,20 +156,22 @@ template ifLogin*(ctx; loggedIn = true; body: untyped) =
     body
   else:
     if loggedIn:
-      respErr "User not logged in. Please signin"
+      respErr needLogin
     else:
-      respErr "User logged in. Please logoff"
+      respErr needLogoff
 
-template minRank*(ctx; minRank: UserRank; body: untyped) =
+template ifMinRank*(ctx; minRank: UserRank; body) =
   ## Runs the code if user have correct rank
   ctx.ifLogin true:
     let
       username = ctx.getSession sess_username
-      user = User.get username
-    if user.rank >= minRank:
+      usr = User.get username
+    if usr.username.len == 0:
+      respErr cannotFindUser
+    elif usr.rank >= minRank:
       body
     else:
-      respErr "Permition denied"
+      respErr rankNotMeet
 
 
 when not defined(windows):
@@ -159,3 +182,64 @@ proc ip*(req: Request): string =
     result = req.nativeRequest.ip
   else:
     result = req.nativeRequest.hostname
+
+from std/json import getStr, getBool, getInt, getFloat
+
+proc updateFields*(
+  obj: var object;
+  node: JsonNode;
+  blacklist: openArray[string] = []
+) =
+  ## Updates the object by node items (if exists)
+  for key, value in obj.fieldPairs:
+    if key notin blacklist:
+      if node.hasKey key:
+        let val = node{key}
+        when value is bool: value = val.getBool
+        elif value is SomeFloat: value = val.getFloat
+        elif value is SomeInteger: value = val.getInt
+        elif value is string: value = val.getStr
+
+
+type
+  ReqDbField = tuple
+    ## Request/DB fields relation
+    req, db: string
+  ReqDbFields = openArray[ReqDbField]
+
+proc getUsing*(
+  table: type;
+  fields: ReqDbFields;
+  node
+): auto =
+  ## Gets a `table` row using some of the provided `fields` in the `node`
+  for (field, inDb) in fields:
+    if node.hasKey field:
+      let val = node{field}.getStr
+      result = table.get(val, [inDb])
+      break
+
+template withUser*(node; usr: untyped; body) =
+  ## Checks if contains user identificators and try to get it, then run the body
+  var fields: seq[string]
+  for (field, inDb) in userIdentificators:
+    fields.add field
+  node.ifContains(atLeast = fields):
+    var usr {.inject.} = User.getUsing(userIdentificators, node)
+    if usr.username.len == 0:
+      respErr cannotFindUser
+    else:
+      body
+
+template withUser*(node; body) =
+  ## Runs `withUser` but with default `usr` variable
+  node.withUser(usr): body
+
+func delInternals*(node: var JsonNode) =
+  ## Deletes all DB internals fields from a JSON node
+  const internalDbPrefixLen = internalDbPrefix.len
+  for key in node.keys:
+    if key.len > internalDbPrefixLen:
+      debugecho key[0..<internalDbPrefixLen]
+      if key[0..<internalDbPrefixLen] == internalDbPrefix:
+        node.delete key